refactor: migrate ENS
