+++ packages/web3-providers/src/types/Web3.ts
@@ -1,6 +1,6 @@
 export namespace Web3BaseAPI {
     export interface Provider<ChainId, Web3> {
-        /** Create an interactable Web3 SDK instance from the given chain ID. */
+        /** Create an interactive Web3 SDK instance from the given chain ID. */
         createSDK(chainId: ChainId): Web3
     }
 }
